[REF] vuestorefront: website_slug, files
Redesigned website_slug as it was clunky and used symbols that are
forbidden on a slugified string.

Also split some code into multiple files, to make it easier to read
and maintain.

Refactored try/except where everything is silently ignored, because that
is a bad practice that can lead to missed errors.