[REF] graphql_vuestorefront -> vuestorefront (#5)
* [REF] graphql_vuestorefront -> vuestorefront

Renaming base vuestorefront module as it is now going to be used as a
fork, with backwards incompatible customizations. Also this name better
matches its purpose.

* [REF] vuestorefront: rem variant_attribute_value_ids

Not used anymore as it had bad performance.

* [REF] vuestorefront: website_slug, files

Redesigned website_slug as it was clunky and used symbols that are
forbidden on a slugified string.

Also split some code into multiple files, to make it easier to read
and maintain.

Refactored try/except where everything is silently ignored, because that
is a bad practice that can lead to missed errors.