[IMP] vuestorefront: better payments support
Redesigned payment acquirers interface to be able to better support
various providers. We now postprocess payment transaction in Odoo model,
so it is easier to extend it.

Also have payment method and payment method issuer objects which are
can be used with payment acquirer if it has these options (must be
implemented for specific acquirer).