[FIX] vuestorefront: odoo_active filter
It makes no sense to filter out payment acquirers on odoo website if
vsf_active is set. Though we still want to have flexibility deciding
if acquirer should be shown in odoo website or/and VSF. So `odoo_active`
field was added that controlls visibility of odoo website side and
`vsf_active` no longer affects it.