[REF] vuestorefront: page render with requests
To make it simpler, we now just use `requests` to make external call,
to get rendered page. Though normally we should still render it directly
instead of relying of external calls.

Previous implementation attempted to render page view directly, but
there are problems with standard odoo where it expects `request` object
to be related with other objects (.e.g. `lang` is expected to be set on
`request`). So we would need to simulate full external call without
actually calling it.